(CPR-345) Making runtime.sh exit 0 if it can't find libdir
If you're not using pl-build-tools, you won't have the need to link in
dynamic libraries and so files, so the runtime.sh script isn't useful.
However, if PROJECT_SHORTNAME is set to `puppet`, which it is by
default, it will expect a puppet directory and when it can't find that
it exits 1, thus killing the build.

This change makes it so that if it can't find the LIBDIR, it just exits
0. This might not be the best change, but trying to not include the
   puppet-runtime component at all was very messy.

Also, puppet-runtime, agent-runtime and runtime.sh all being things I
have to keep context on in my head is just confusing.